% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{dataframe}

\item{init_node}{number for starting node}
}
\value{
shortest path between the selected node and all other node
}
\description{
calculates the shortest path between every nodes in a graph
}
\references{
(https://en.wikipedia.org/wiki/Dijkstra%27s algorithm)
}
